pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = ""
authors = [
    {name = "Aarav Asthana",email = "aaravasthana2308@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
[tool.poetry.dependencies]

python-dotenv = "^1.1.0"
google-generativeai = "^0.8.5"
pymupdf = "^1.26.1"
pdfplumber = "^0.11.7"
pytesseract = "^0.3.13"
pillow = "^11.2.1"
spacy = "^3.8.7"
requests = "^2.32.4"
streamlit = "^1.45.1"
openpyxl = "^3.1.5"
matplotlib = "^3.10.3"
plotly = "^6.1.2"




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
